+++ src/pub/overview.js
@@ -30,7 +30,7 @@ class Overview extends Pure_component {
         this.update_window_dimensions();
         window.addEventListener('resize', this.update_window_dimensions);
     }
-    componentWillUnmount(){
+    willUnmount(){
         window.removeEventListener('resize', this.update_window_dimensions);
     }
     render(){