+++ util/http_hdr.js
@@ -148,6 +148,23 @@ E.browser_default_header_order_http2 = function(browser){
     return headers[browser]||headers.chrome;
 };
 
+// reverse pseudo headers (e.g. :method) because nodejs reverse it
+// before send to server
+// https://github.com/nodejs/node/blob/v12.x/lib/internal/http2/util.js#L473
+function reverse_http2_pseudo_headers_order(headers){
+  let pseudo = {};
+  let other = Object.keys(headers).reduce((r, h)=>{
+      if (h[0]==':')
+          pseudo[h] = headers[h];
+      else
+          r[h] = headers[h];
+      return r;
+  }, {});
+  pseudo = Object.keys(pseudo).reverse()
+      .reduce((r, h)=>{ r[h] = pseudo[h]; return r; }, {});
+  return Object.assign(pseudo, other);
+}
+
 E.like_browser_case_and_order_http2 = function(headers, browser){
     let ordered_headers = {};
     let header_keys = E.browser_default_header_order_http2(browser);
@@ -164,7 +181,7 @@ E.like_browser_case_and_order_http2 = function(headers, browser){
         if (!header_keys.includes(h))
            ordered_headers[h] = req_headers[h];
     }
-    return ordered_headers;
+    return reverse_http2_pseudo_headers_order(ordered_headers);
 };
 
 let parser = new HTTPParser(HTTPParser.REQUEST), parser_usages = 0;